Unify import strategy (#1975)
Unify the strategy for handling additional imports in the printing stage
for different languages. This allows `get_additional_imports` and
`add_import` to be moved to the superclass
`codegen.printing.CodePrinter` and will make it easier to handle
different import strategies (e.g. #1657)

**Commit Summary**
- Move `get_additional_imports` to `codegen.printing.CodePrinter`
- Move `add_import` to `codegen.printing.CodePrinter`
- Use a dictionary for imports in `codegen.printing.FCodePrinter` to
match what is done in `codegen.printing.CCodePrinter`
- Change the storage type of the `_additional_imports` dict in
`codegen.printing.PythonCodePrinter` to match what is done in
`codegen.printing.CCodePrinter`