Remove use of ndarrays in Fortran array argument wrapping (#1963)
Remove the use of `ndarray_t` in the C wrapper of a function translated
to Fortran. Fixes #1962 , making it easier to remove `ndarray_t` later
and reducing complexity.

The functions `_wrap_FunctionDefArgument` and
`_wrap_BindCFunctionDefArgument` are modified to return a dictionary
with the useful information that they calculate. In addition to the body
returned previously the function arguments are additionally returned.
This makes it easier to pass multiple variables to the function and
simplifies `_wrap_FunctionDef` slightly

**Commit Summary**
- Add the function `get_strides_and_shape_from_numpy_array` to calculate
the strides and shape of a NumPy array
- Use `var` to avoid the repetition of `expr.variable` in
`CCodePrinter_print_Declare`
- Add code to declare a C stack array
- Use consistent naming in `c_to_python_wrapper.py`. The original name
is now always the name that is used to recognise a variable in the scope
- Add `_get_array_parts` to get the data, shape, and strides of an array
- Modify `_wrap_FunctionArgument` and `_wrap_BindCFunctionArgument` to
return dictionaries for readability. The arguments passed to the
function are also returned so operations such as adding `ObjectAddress`
can be done in the relevant function instead of being added a posteriori
in `_wrap_FunctionDef`