Correct Numpy version check (#1992)
Following #1988 the tests are failing on MacOS
(https://github.com/pyccel/pyccel/actions/runs/11013940047). This is due
to a change that was attributed to NumPy v2.0 but actually only occured
in NumPy v2.1:
https://numpy.org/devdocs/release/2.1.0-notes.html#changes

![image](https://github.com/user-attachments/assets/3dbefb6a-58f4-45bb-9bd1-306f2f4e9e13)

This PR corrects the version check.

Additionally it excludes v2.0.* from the dependencies as the handling of
integers on Windows seems to be broken with this version.