Fix type bugs with NumPy v2 (#1994)
After NumPy v2 the NumPy precision is preferred over the precision
provided by any native Python types involved in calculations. This
causes the result types of arithmetic operators to change. This was not
a problem in the devel branch due to our use of `np.result_type` to
calculate the types using information about both the type and the
precision.

This PR fixes the problem to prepare the patch release. It does this by
ignoring native types. The fix handles most cases but is not perfect. It
is very difficult to do better without making the changes in #1756.