Fix test failures with Intel compiler (#1999)
Fix the Intel tests in the release branch. In this branch the only
errors are due to a compiler bug. The bug has been reported
[here](https://community.intel.com/t5/Intel-Fortran-Compiler/Compiler-Bug-cmplx-does-not-handle-Y-argument-correctly/m-p/1634388)
but a simple workaround has been implemented. This therefore fixes #1932.

Reactivate some skipped Windows tests which should have been
reactivated 2 years ago (#735 was fixed in January 2022).