Add Fortran support for list declarations (#1976)
Add support for list declarations. List initialisation is left for a
second PR. This PR simply ensures that a list object can be declared and
that the gFTL import is correctly included and linked during the
compilation. This PR does not fully handle lists of classes. Support for
indexing is also added in order to check that the contents of the list
are correctly initialised.

**Commit Summary**
- Correct module docstring of `pyccel.ast.c_concepts`
- Add `is_external` property to `Module` to be able to create modules
which are defined inside `#include` statements (this is how gFTL works)
- Preserve the ordering of imports in Modules
- Add a `KindSpecification` class to be able to print statements such as
`(kind=i64)`
- Add an `IteratorType` class to describe the type of a container
iterator (e.g. `Vector_int_Iterator` in gFTL and `vec_int_iter` in STC)
- Add a `MacroDefinition` class to describe the declaration of macros
with `#define`
- Save literal containers used in NumPy in tuples not lists to avoid
printing with gFTL
- Ensure extension modules can be generated (e.g. a module for each gFTL
type)
- Allow `FCodePrinter.print_kind` to take a type
- Add an `_apply_cast` function to avoid code duplication and ensure
that casts applied to lists/sets are applied to their members.
- Add a `_get_node_without_gFTL` to allow printing literals for
initialisation or function calls (e.g `[1,2,3]` instead of
`Vector([1,2,3])`)
- Fix basic allocation of containers
- Fix basic deallocation of containers
- Add basic indexing of lists (for tests)
- Add `test_sum_bool` to increase coverage
- Skip tests which initialise arrays from lists stored in variables
- Skip tests using list comprehension
- Skip tests which copy lists
- Skip tests which slice lists
- Activate list declaration tests