Fix the worst Intel bugs (#2002)
This PR incorporate #1999 into the `devel` branch. #1999 included a
workaround for an intel compiler bug.
In the `devel` branch the intel branch wasn't working due to #2001 .
This bug is not present in the `main` branch as there is a memory leak
when returning arrays from functions (these arrays are never freed). The
fix for this is non-standard and does not work with the Intel compiler.
This is documented in #2001 however, for now, to prevent further Intel
bugs being introduced without us noticing, the call to `free` is
deactivated for the Intel compiler. This allows the tests to pass