Allow homogeneous tuples as function arguments (#1850)
Allow 1D homogeneous tuples as function arguments. Fixes #738.
Homogeneous tuples as arguments already worked in the low level code as
homogeneous tuples are handled as arrays, however the wrapper was
missing which meant that such functions couldn't be called from Python.
In order to handle multi-level tuples some reorganisation of the wrapper
was necessary. Namely the introduction of the
`_extract_X_FunctionDefArgument` functions. However Python multi-level
tuples may have different lengths in the elements. It proved difficult
to check the size of the elements in a generic way so this was left for
later work (the obvious solution is to use lists to support this).

**Commit Summary**
- Update docs
- Update `BindCFunctionDefArgument` to handle tuples as strideless
arrays
- Add descriptions of more tuple functions from `Python.h`. Group these
functions together in `ast.cwrapper`
- Make `dtype` a searchable attribute of `SyntacticTypeAnnotation`
- Fix filter for used template type annotations so it also searches
through indexed elements to find objects in a tuple annotation.
- Add tests for tuples as arguments
- Rename `_get_check_function`->`_get_type_check_condition`.
- Allow `_get_type_check_condition` to return a condition which is not a
function call (e.g. a Variable)
- Add `body` argument to `_get_type_check_condition` to allow additional
commands to be run before determining if the type is correct.
- Add type checks for homogeneous tuples
- Introduce `_extract_X_FunctionDefArgument` functions to describe
container unpacking recursively
- Add `_extract_HomogeneousTupleType_FunctionDefArgument` function to
unpack 1D homogeneous tuples.
- Ensure errors are raised for multi-D tuples.