Add support for @property (#2089)
The semantic handling and the printing of methods using the `@property`
decorator was already working, however official support for the
decorator was missing due to the missing wrapper support. This PR adds
the wrapping and removes the warning about this decorator. Fixes #1834

**Commit Summary**
- Add documentation
- Prefer assertions for errors that only affect developers
- Allow the setter of a property to be missing (it is then saved as
`None`)
- Add wrapping of a class property. The arguments of such functions are
slightly different to those of a basic function.
- Add `-Wno-incompatible-pointer-types` to Python flags to remove
unavoidable warnings
- Add tests