Change the order of the arguments of `FunctionDef` (#2101)
In order to fix #337 the handling of function results must be changed.
Specifically, the handling of the case where nothing is returned is
non-trivial. It is therefore better to handle this with a default value,
however adding a default value requires the order of the arguments of
the `FunctionDef` class to be changed. This PR makes that change in
preparation