Fix integer handling with NumPy 2.0 on Windows (#2116)
Fix integer handling with NumPy 2.0 on Windows. The default integer is
correctly deduced however the handling in the wrapper was incorrect.
This was fixed (fixes #2115). Additionally the dtype error output is
cleaned up so the positions of the datatypes in NumPy's enum are not
hard coded on our side.

This fix was complicated by the fact that `np.random.randint`'s default
dtype is maintained at `np.long` and therefore does not match the
default used by all other NumPy functions in v2:

![image](https://github.com/user-attachments/assets/3f331352-2e92-44d4-8176-1e410af5cfd6)