Issue #983 | Add Round function and implement the conversion to C, Fortran, and Python (#996)
Add support for Python's `round` function which is implemented with a
banker's round. An integer and a floating point implementation are both
required to avoid floating point errors when handling integers. Fixes
#983.

A minor bug fix is also included to ensure that the error mode is set
back to 'user' when nothing is specified if the developer mode was used
previously.

---------

Co-authored-by: Nandini Raja <nandini.raja@thoughtworks.com>
Co-authored-by: Emily.Bourne <louise.bourne@gmail.com>