Fix bug in `get_strides_and_shape_from_numpy_array` (#2138)
Fix bug in `get_strides_and_shape_from_numpy_array`. Fixes #2137 . The
bug was occurring for arrays with rank > 1 with C-ordering. The fix is
simply to correct the order in which we iterate over the elements.

In order to more easily detect this kind of error the `--debug` flag (or
`debug=True` for epyccel) is automatically added in the pull request
tests via an environment variable `DEBUG`.

Debug mode has highlighted extra errors, most of which were already
detected. See #1668 and #2042. Unravelling is now tested more explicitly
by returning arrays instead of only returning their shape.