Remove BindCFunctionDefResult (#2157)
Remove `BindCFunctionDefResult`. It is replaced with the class
`BindCVariable` which links new C-compatible variables to their
equivalent in Fortran. When returning arrays multiple variables are
returned from Fortran this is now represented using tuples. The type
`BindCArrayType` subclasses `InhomogeneousTupleType` to describe the
tuple `(a.data, *a.shape)` for an array `a`.

The code in `fortran_to_c_wrapper.py` is re-written to use
`_extract_X_FunctionDefResult` functions to improve readability.

This PR is important as a first step towards #337 . After #337 we will
have 1 `FunctionDefResult` per function so we cannot use
`FunctionDefResult` or its subclass `BindCFunctionDefResult` to describe
the relationships between array data and shape.

This PR will also make it easier to implement #2060 .

**Commit Summary**
- Make docstrings optional for `_extract_X_FunctionDefResult` to avoid
copy/paste docstrings
- Add new datatype class `BindCArrayType`.
- Rename existing class `BindCVariable`->`BindCModuleVariable`
- Add new class `BindCVariable`
- Remove class `BindCFunctionDefResult`
- Store `FunctionDefResult(BindCResultVariable)` in `BindCFunctionDef`
instead of `BindCFunctionDefResult`
- Add function `flatten_tuple_var` to obtain all low-level variables in
a multi-level tuple.
- Preserve scope when creating external function definitions so
inhomogeneous tuple elements can be retrieved for arguments or results.
- Add method `_extract_BindCArrayType_FunctionDefResult` to
`CToPythonWrapper` as Fortran and C NDArray wrapping is quite different.
- Introduce `_extract_FunctionDefResult` methods in `FortranToCWrapper`
to improve readability.
- Add `_get_bind_c_array` method in `FortranToCWrapper` to avoid code
duplication.