Fix shape of multi-level containers (#2180)
Fix the shape of multi-level containers. The shape should be `None` for
0-d objects. Otherwise it should be a tuple. The length of this tuple
should be equal to the `container_rank` of the class_type being
described. Homogeneous tuples are an exception to this rule. As they are
saved as NumPy ND arrays they need to have a shape equivalent to an ND
array shape. Correcting the shape means that the shape of elements of a
list are not saved. They are unknown and are set to `PyccelArrayShape`
when the list is indexed. Fixes #2175

**Commit Summary**
- Add `__epyccel__*` to gitignore to ignore files generated when running
pytest in parallel.
- Ensure all 0D objects (including classes) have a shape of `None`
- Update the shape of a tuple calculated in `PythonTuple`
- Update the shape of a `PythonList` to be of rank 1
- Add a `shape_is_compatible` method to `DataType` to check if the shape
has the expected format
- Change inheritance of `CustomDataType` as it is not usually a
"Container" in the sense understood by Pyccel.
- Add a function `get_shape_of_multi_level_container` to calculate the
shape of a NumPy array from a list of lists or tuple of tuples
- Correct the shape of an `IndexedElement` using
`PyccelArrayShapeElement` to get the shape of elements of unknown shape