Clean up Fortran-to-C argument wrapping (#2232)
Clean up Fortran-to-C argument wrapping to match the patterns used in
the Fortran wrapper for results, and in the C wrapper for both results
and arguments. This will make it much simpler to add support for
sets/dicts/strs as arguments.

**Commit Summary**
- Add `_extract_.*_FunctionDefArgument` to the list of functions which
all behave the same and therefore do not require duplicated
documentation. The documentation for this function is found in
`_extract_FunctionDefArgument`
- Update developer docs to describe the use of
`_extract_FunctionDefArgument`
- Stop differentiating between `BindCFunctionDef.arguments` and
`BindCFunctionDef.bind_c_arguments`. Instead save the argument variables
in a `BindCVariable` describing both the original variable and the new
c-compatible variable.
- Handle printing inhomogeneous tuples as C arguments (wrapping is not
added) to allow the printing of a `BindCVariable` representing an array
- Add `_extract_.*_FunctionDefArgument` methods in
`fortran_to_c_wrapper` to handle different types with different code.
This simplifies `_get_function_def_body`
- Correct variables incorrectly marked with `is_argument` due to clones.
- Remove the now unused class `BindCFunctionDefArgument`
- Remove `flatten_tuple_var` whose functionalities are completely
handled by `Scope.collect_all_tuple_elements`
- Remove `FortranWrapper._get_call_argument`. Its functionalities are
moved to `_extract_.*_FunctionDefArgument` to make it easier to define
them differently for different types.