Fix shape type error (#2245)
The shape of an `IndexedElement` was not always saved in a
`TypedAstNode`. This causes problems as the `_shape` is an
`attribute_node`. The reason for this error was because the construction
of `_shape` is completed after the call to `super().__init__` which
auto-converts integers to LiteralIntegers. This error is now fixed. The
`LIST_OF_TUPLES` error check which was reporting this error had been
erroneously moved to the blocking folder. It is now moved back