Improve issue templates  (#2255)
Fixes #2254.

Update the issue templates in folder `.github/ISSUE_TEMPLATE/`. The
changes involve the replacement of the existing Markdown issue templates
with new YAML-based issue templates. These new templates provide a more
structured and user-friendly experience for reporting issues, suggesting
features, and other contributions. For convenience, we provide links to our
Contributing Guidelines, CHANGELOG, and Q&A Discussions.

In addition to the existing templates, a new template is added for
suggesting improvements.