Provide context clues to epyccel to access constants and modules (#2279)
Use `inspect` to provide context clues to epyccel. This gives Pyccel
access to the calling context. This is used to insert constants (literal
scalars or TypeAliases) and known modules. This will allow us to
deprecate the `@template` decorator and stop writing `from numpy import np`
in almost all test functions despite the module being available at
the top of the file.

**Commit Summary**

- Allow NumPy literals to be used to create `Literal` instances
- Allow a `VariableTypeAnnotation` to be passed to `process_dtype`
- Ensure `Literal` objects contain pure Python literals of the correct type
- Add `context_dict` argument to functions to allow context information
to be provided to the semantic parser
- Use evaluated function signature when printing functions in epyccel to
ensure types are known
- Rename `function_or_module` in `epyccel` and `epyccel_seq` to reflect
the fact that these methods handle classes too
- Use `context_dict` to insert literals, recognised imports and typehints
- Add tests