Fix LAPACK installation on windows (#2294)
The Windows CI is failing as the University of Tennessee has changed the
permissions on its website. It is therefore no longer possible to
download the compiled dll files from this site.

This PR changes the installation process to collect LAPACK from the
GitHub reference repository. It is compiled with CMake and the result is
cached to avoid recompiling every time we use a Windows runner. This new
installation fixes #1091