Do not install optional dependencies:
Reading the list of [test] optional dependencies from the pyproject.toml
file is not feasible for standard installs, because this file is not
installed. Providing this feature for editable installs only seems
pointless, hence we remove it altogether.

Instead, we print a clear error message which describes how to install
Pyccel with its optional [test] dependencies.