Fix memory handling for lists of lists (#2270)
Generate the necessary memory handling in the semantic stage to
correctly deallocate multi-dimensional objects. Add C printing to add
prelimary tests for this. Lists of lists are now supported in C. Fixes
#1945. Fortran support is left for a later PR. Each operator needs some
tweaking in the output to handle managed memory. This PR only does this
for `pop`, `append` and `popitem` as a proof of concept. Similarly while
lists of lists/sets/dicts are implemented more complex element types
(e.g arrays/classes) are left for a later PR

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>