Fix integer linspace (#2312)
Fix wrong result when using integer dtype in `np.linspace`. The
difference between our previous implementation and NumPy's
implementation is that we were relying on `int` to do the rounding,
while NumPy were using `floor`. This is fixed for both C and Fortran.
Fixes #2126 . The code is further cleaned up to use f-strings instead of
relying on a string template and calling `format` as a last step.