Fix missing index in loop unravelling for loop of length 1 (#2315)
Fix missing index in loop unravelling for loop of length 1. This is a
general fix which works but leads to expressions such as `0+4`. Fixes
#2042.

Use `LiteralInteger(0)` as index when size is known to be 0 to ensure
better simplification. I am not 100% certain that this includes all
locations where indices may be generated so I have also left the general
fix in place.

Pre-existing tests which fail consistently in debug mode (now used in
the CI) are reactivated.