Fix class scoping (#2314)
Fix class scoping. Methods should have the Module as `parent_scope`
(used for name clashes), but the resulting function should be saved into
the scope of the class. This can cause problems for retrieving the name
of the function itself so the syntactic<->semantic mapping of the name
of a function is added to the scope of that function in the
`create_new_function_scope` in the semantic stage. Fixes #2093

**Commit Summary**
- Add assertions to `FunctionDef` and `ClassDef` to ensure the scopes
are well-formed
- Exit class scopes before visiting methods so they have the correct
parent
- Correct size of homogeneous tuple in wrapper (not clear why this error
became apparent in this PR)
- Rename `self._current_function`->`self._current_function_name`
- Use a list to store `Parser._current_function_name` to ensure history
can be cleanly retrieved
- When a function is inserted into the scope, use it to update a new
variable `self._current_function` (useful to avoid searching for the
function in the scope, which is difficult for a class method where the
scope is not readily available)
- Move `Parser._annotate_the_called_function_def` from `Parser` to
children `SyntacticParser` and `SemanticParser` to allow a more complex
implementation for `SemanticParser`
- Use 2 name arguments for
`SemanticParser._annotate_the_called_function_def` to ensure correct
handling of syntactic and semantic names
- Call correct scope from `collect_tuple_element` if the argument is a
`DottedVariable`
- Remove unused code
- Use new `current_function_name` attribute to access
`self._current_function_name` to avoid code duplication
- Change the arguments of
`SemanticParser._annotate_the_called_function_def` to make the
`function_call_args` argument compulsory. This reduces code duplication
- Insert functions into the correct scope
- Add a test (from the issue) with a class with an attribute whose name
conflicts with a variable found in the other methods of the class