Fix visitation order for classes (#2316)
Fix the visitation order for class methods. Now `_visit_ClassDef` only
adds the syntactic methods to the scope. This matches what is done for
functions in `_visit_Module`. When a method is needed it is visited. In
addition if a class instance is passed as an argument, the class's
`__init__` method is visited if it has not already been. This ensures
that the attributes are available. Finally the destructor is created in
`_create_class_destructor` to ensure memory safety. Fixes #1814 .
This means that `inline` class methods are now handled like normal class
methods. The implementation from #1728 is used, so inline class methods
are no longer exposed to Python in the interface.

**Commit Summary**
- Remove dead code in `pyccel.ast.core.ClassDef`
- Only remove syntactic methods from a ClassDef interface if they are
present. As methods are now visited out of order one implementation of
an interface may be visited (removing the syntactic node), independently
of the other implementations (for whom the syntactic node will already
have been removed).
- Stop printing inline methods in generated code. This matches the
documented behaviour and is necessary as inline methods may not have
been visited in the semantic stage.
- Split `_visit_ClassDef` into a step before visiting a module body
(`_visit_ClassDef`) and a step after visiting the body and the class
methods (`_create_class_destructor`) to add any necessary deallocations
- Ensure class methods are annotated when they are first used
- Move inline handling to inside `_annotate_the_called_function_def` to
reduce code duplication
- Fix position of call to garbage collection in `_visit_FunctionDef`
- Move the creation of an empty `__init__` method to the syntactic stage
- Add tests
- Add a test for the broken case seen in #2310 
- Stop calling inline methods from the Python side of tests

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>