Fix handling of optional arguments in conditions (#2319)
Fix handling of optional arguments in conditions to prevent seg faults
when the lazy evaluation assumption is not valid. Fixes #1668
Expressions such as `if a and ...` are converted to `if a:\n if ...`.
More complex expressions raise errors asking users to reorganise their
code.

**Commit Summary**
- Remove cast to `PythonBool` in `IfSection.__init__` (replaced with
assertion that argument is a boolean).
- Change condition used for `IfSection`s to explicitly obtain a boolean
- Add `set_current_ast` to `If` which passes the AST down to the
`IfSection`s
- Consolidate recursive `PyccelAnd` and `PyccelOr` objects into 1 object
to help with special treatment
- Ensure arguments to `PyccelAnd` and `PyccelOr` are booleans
- Add recursive `treat_condition` function in `_visit_IfSection` to fix
handling of optional arguments
- Correct output type for `PyUnicode_Check`