Fix printing of Python interfaces (#2320)
Add if blocks in Python output to ensure support for implementations
that differ for different types. Fixes #802.
Reactivating the tests showed a problem with the casting of NumPy
arrays. This is also fixed.
In order to print the conditional interfaces support is added for
`numpy.ndarray`.

These changes are required so inline interfaces can be printed in .pyi
files.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>