Parse stub files to accelerate compilation (#2317)
Parse `.pyi` files instead of `.py` files when importing files to avoid
unnecessarily parsing complex/costly body expressions. Fixes #297

**Commit Summary**
- Fix incorrect condition in pylint CI
- Ensure pre-existing `Interface`s are removed when they are replaced by
a new `Interface` with an extra element
- Add support for `typing.overload`
- Use `pathlib` instead of `os.path` in files where this is simple:
    - pyccel/codegen/compiling/basic.py
    - pyccel/codegen/pipeline.py
- Move `typing.overload` import in pycode to avoid unnecessary imports
- Print an `Interface` instead of multiple functions with `@overload`
for Python to enable treating at semantic using existing code.
- Fix overloaded function name (should match original Python code)
- Use `self._is_header` to preserve the names of function arguments
- Ensure module init is recognisable
- Add two extra arguments to `Parser`:
- `output_folder` : This is important to be able to create paths
relative to the correct folders
- `original_filename` : This is important to track both the original
Python file **and** the `.pyi` header
- Update `get_filename_from_import` to use `pathlib` and to ensure all
returns are at the bottom of the function to avoid duplicating the logic
deciding which file should be chosen
- Update the definition of "header file" to include `.pyi` files
- Update semantic parsing to handle the fact that an `Interface` can be
spread over multiple `FunctionDef` objects when the `@overload` operator
is in play
- Correct result type
- Delete import tests which don't treat files being imported. Raises an
error as file to be imported has not been translated