Deprecate `@template` decorator (#2331)
Deprecate the `@template` decorator. We can fully rely on `TypeVar` and
`UnionTypeAnnotation` to deduce the functions in the interface. This is
one of the main steps towards #1487 .
Add a `context_dict` argument to `epyccel` for passing non-global
`typing.TypeVar` objects.

**Commit Summary**
- Delete obsolete code related to templates
- Rename `TypeVar.name`->`TypeVar.name_str` to keep
`PyccelFunction.name` for Python printing
- Ensure `TypeVar` declarations appear in Python translations
- Prefer `Final[X]` over `const X` as it respects Python syntax for
forward declarations
- Add a `context_dict` argument to `epyccel` to allow non-global
`typing.TypeVar` objects to be used via the context
- Remove `templates` argument from `lambdify` (`context_dict` does the
same job)
- Remove the `@template` decorator
- Remove the `#$ header template` syntax
- Update `_visit_FunctionDef` to calculate `Interface`s using logic
based on `TypeVar`s and `UnionType`s instead of templates