Improve inline handling (#2330)
Inline functions are not wrapped. This means that they do not need to be
handled in the same way as normal functions. Up to now they have been
handled in the same way as other functions in the semantic stage then
the codegen does some error-prone calculations to swap in the arguments
passed to the inline function. After this PR inline methods are handled
in the semantic stage. They are never translated. Rather the arguments
and results of the function call are swapped in and the resulting body
is appended to the body of the calling scope. This is less error-prone
as there is no risk of the syntactic node changing while the semantic
node is being constructed.
This handling means that type annotations are not compulsory on
`@inline` methods.

This is a critical step towards eliminating macros in Pyccel as
`@inline` functions (when working correctly) are a drop-in replacement.
The removal of the obligation to add type annotations will also help for
wrapping scipy functions where the types can depend on the function
arguments.

Fixes #2322. Fixes #2043 

**Commit Summary**
- Avoid invalidating syntactic nodes that were converted to semantic
nodes
- Ensure `FunctionDef.pyccel_staging` matches `FunctionDef.is_semantic`
- Allow `PyccelSymbol` to be found inside an `AnnotatedPyccelSymbol`
- Remove C and Fortran implementations of inline function handling
- Ensure inline methods are not printed in Fortran
- Simplify Python printing of inline functions
- Add missing import of `inline` decorator 
- Ensure module initialisation function is not called in Python even if
it is renamed during the import
- Allow `remove_variable` to be called from a scope which can access the
variable but does not own the declaration (e.g. when removing a variable
created for an inline call from a loop scope)
- Add `Scope.symbol_in_use` function
- Translate inline function bodies inline, when calling an inline method
- Remove dead test (inline functions no longer check their argument
type)
- Increase coverage in tests
- Reactivate test due to bad unravelling

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>