Fix missing import (#2351)
When creating the release branch I noticed an import which was
accidentally deleted although it is used in the file. This PR puts it
back. The object is used here:
https://github.com/pyccel/pyccel/blob/devel/pyccel/codegen/printing/ccode.py#L2795

This is the fallback case for Constant which we do not use currently
(i.e. it is not currently tested). It will either be removed with
https://github.com/pyccel/pyccel/pull/1811 or will start being used and
tested