Fix Import regression (#2354)
Previously imports like `import a.b.c as d` were saved as `d` however
this makes it difficult to reuse parsing when 2 files use `a.b.c` with
different names. This was therefore changed so the key is now `a.b.c`.
The case `import a.b.c as d` was correctly handled, but `import a.b.c as
c` was not as it was previously a special case. The special case has now
been corrected and a test is added. Fixes #2353.