Clean up `__all__` usage and CI (#2369)
One of the complexities of the CI comes from the fact that a `base`
argument is required to compare the current code to the code on the
`devel` branch. This is done to filter the errors so devs opening a new
PR are not responsible for technical debt. This PR cleans up the
technical debt associated with the `pyccel_lint` test. As part of this
the CI interface is cleaned up to make it easier to call it on a local
computer.

**Commit Summary**

- Rename
`ci_tools/check_slots.py`->`ci_tools/check_pyccel_conventions.py`
- Remove `base` argument to `pyccel_lint` workflow
- Remove required `output_file` argument from
`ci_tools/check_pyccel_conventions.py`,
`ci_tools/check_pylint_commands.py` and
`ci_tools/check_python_capitalisation.py` to display output on command
line for devs
- Modify `ci_tools/check_python_capitalisation.py` to calculate the
files to be checked in the test itself
- Improve alphabetisation check in
`ci_tools/check_pyccel_conventions.py`. Previously the `__all__`
variable should be entirely alphabetised. Now it should be alphabetised
by group of code concepts. This allows comments to be used to break the
`__all__` into relevant sections (e.g. base classes, exposed classes)
- Clean up docstrings in CI
- Remove `file_changed` check from `ci_tools/check_pylint_commands.py`
as no errors are raised on the `devel` branch
- Add developer testing docs to describe the details of the CI docs,
which can be run locally and how to run them
- Alphabetise remaining `__all__` variables