Fix links in docs (#2379)
Fix links in docs. Absolute links should be used in `README.md` as this
file is used as-is on PyPI. Everywhere else relative links should be
used. These are resolved to url links by sphinx.

The list of documentation pages is removed from the README as it is now
easily locatable on the official docs page. "Getting Help" sections at
the end of problem-specific documentation pages are removed. References
to the email address `<pyccel@googlegroups.com>` (which is never used
and probably sends emails to the wrong people) are removed.